Fixed update method and clean up files
