+++ app/javascript/packs/application.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { render } from 'react-dom'
 
-import App from '../react/components/App'
+import App from '../react/App'
 import RedBox from 'redbox-react'
 
 document.addEventListener('DOMContentLoaded', () => {
