+++ app/javascript/react/ui/AddToLibraryButton.js
