+++ app/javascript/react/ui/OpenExchangeButton.js
